New tabs and Fixed ones
Minecraft is now a tab. I need to look for people to write the paragraphs but the code is ready. Papa's sushiria has been worked on and improved, there is still some work to do there. There havce also been some minor changes to the index, and the about tab is now written. Improvements coming soon!