New Tab!
Fortnite is now a tab. The next step is to find an author for this tab as well as to create other tabs and clean up the fonts and the style on completed tabs. I have also found an author for the minecraft tab, and it will be completed by the end of this next sprint.